+++ src/App.tsx
@@ -3,11 +3,10 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 import Landing from "./pages/Landing/Landing";
 import LogIn from "./pages/LogIn/LogIn";
-import Error from "./pages/error/Error";
+import Error from "./pages/Error/Error";
 import SignUp from "./pages/SignUp/SignUp";
-import SignComplate from "./pages/signComplate/SignComplate";
+import SignComplate from "./pages/SignComplete/SignComplete";
 import Interest from "./pages/Interest/Interest";
-
 
 function App() {
   return (
